Добавил react-router. Подрефакторил компоненты
