+++ src/main/resources/ui/js/containers/Application/Application.js
@@ -1,14 +1,20 @@
 import React from "react";
 import {Provider} from "react-redux";
 import store from "../../store"
-import RegistrationForm from "./components/RegistrationForm";
+import RegistrationForm from "../RegistrationForm"
+import {BrowserRouter, Route, Link} from "react-router-dom";
+import AuthForm from "../AuthForm";
 
 class Application extends React.Component{
     render(){
         return (
             <Provider store={store}>
-                <h1>Форма регистрации</h1>
-                <RegistrationForm />
+                <BrowserRouter>
+                    <Link to={"/"}>Вход</Link>
+                    <Link to={"/registration"}>Регистрация</Link>
+                    <Route path={"/"} exect component={AuthForm}/>
+                    <Route path={"/registration"}  component={RegistrationForm}/>
+                </BrowserRouter>
             </Provider>
         )
     }
