+++ src/main/resources/ui/js/containers/Application/index.js
@@ -1,0 +1,3 @@
+import Application from "./Application";
+
+export default Application;
