+++ src/main/resources/ui/js/containers/AuthForm/AuthForm.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+class AuthForm extends React.Component {
+    render(){
+        return (
+            <h1>Вход</h1>
+        )
+    }
+}
+
+export default AuthForm;
