+++ src/main/resources/ui/js/containers/RegistrationForm/index.js
