+++ src/main/resources/ui/js/index.js
@@ -1,5 +1,5 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import Application from "./containers/Application/Application";
+import Application from "./containers/Application";
 
 ReactDOM.render(<Application />, document.getElementById("app"));
